interval start test

+++ hash_pool_enty2.c
@@ -235,7 +235,7 @@ int main()
     printf("\nenty_ar_chunk_storage is open\n");
     fseek(chunk_storage_file, 0, SEEK_END);
     long nLen1 = ftell(chunk_storage_file);
-    for(int i=0 ; i<nLen1; i++) {
+    for(long i = 0 ; i < nLen1; i++) {
         unsigned char EndIntervalStart[2];
         fseek(chunk_storage_file, i, SEEK_SET);  // read from position
         fread(EndIntervalStart, 1, 1, chunk_storage_file);
